Extract core functionality to HockeyAppAndroidEngine class Leaving HockeyAppAndroid only as a bridge between the MonoBehaviour Life cycle and ability to start coroutines. This enables more flexible integration into apps where HockeyApp configuration is less static.
